modified dashboard submission_admin and header portal

+++ application/controllers/Dashboard.php
@@ -1060,20 +1060,29 @@ class Dashboard extends CI_Controller {
 	}
 
 	public function saveapprovaluser(){
-		$q1 = "delete from t_adminuser where id_projek = ".$_POST['id_project'];
-		$q1 = $this->db->query($q1);
-
-		$this->db->query("UPDATE t_projekprofil SET status = 6, komen1 = '".$_POST['komen1']."' WHERE id_projek = ".$_POST['id_project']);
-		foreach ($_POST['username'] as $key => $value) {
-			$this->db->query("insert into t_adminuser (id_projek, username) values (".$_POST['id_project'].", '".$value."')");
+		if($_POST['komen1'] == '' || $_POST['komen1'] == null){
+			return false;
+		}else{
+			$q1 = "delete from t_adminuser where id_projek = ".$_POST['id_project'];
+			$q1 = $this->db->query($q1);
+	
+			if($_POST['xxi_xxi'] == 4){
+				$this->db->query("UPDATE t_projekprofil SET status = 7, komen1 = '".$_POST['komen1']."' WHERE id_projek = ".$_POST['id_project']);
+				foreach ($_POST['username'] as $key => $value) {
+					$this->db->query("insert into t_adminuser (id_projek, username) values (".$_POST['id_project'].", '".$value."')");
+				}
+			}else{
+				$this->db->query("UPDATE t_projekprofil SET status = 6, komen1 = '".$_POST['komen1']."' WHERE id_projek = ".$_POST['id_project']);
+				foreach ($_POST['username'] as $key => $value) {
+					$this->db->query("insert into t_adminuser (id_projek, username) values (".$_POST['id_project'].", '".$value."')");
+				}
+			}
+			return true;
 		}
-
 		// $this->db->query("UPDATE t_projekprofil SET status = 7, komen1 = '".$_POST['komen1']."' WHERE id_projek = ".$_POST['id_project']);
 		// foreach ($_POST['username'] as $key => $value) {
 		// 	$this->db->query("insert into t_adminuser (id_projek, username) values (".$_POST['id_project'].", '".$value."')");
 		// }
-
-		return true;
 	}
 
 	public function backtoadmin(){
